# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hajj Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new pilgrim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Phone or passport already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone/passport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # GROUPS
  /groups:
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join a group by invite code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGroupInput"
      responses:
        "200":
          description: Joined group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{id}/members:
    get:
      operationId: getGroupMembers
      tags: [groups]
      summary: List group members
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMember"

  # LOCATIONS
  /locations/sync:
    post:
      operationId: syncLocations
      tags: [locations]
      summary: Sync batch of location points (offline-safe)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationBatch"
      responses:
        "200":
          description: Synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
...
  /manasik:
    get:
      operationId: getManasik
      tags: [manasik]
      summary: List all Hajj rituals with user progress
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Manasik list with progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManasikItem"
  /manasik/{key}/progress:
    patch:
      operationId: updateManasikProgress
      tags: [manasik]
      summary: Update progress for a mansak ritual
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManasikProgressUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManasikItem"

  # PERSONAL PLACES
...
  /emergency/sos:
    post:
      operationId: sendSos
      tags: [emergency]
      summary: Send an SOS emergency alert
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SosInput"
      responses:
        "201":
          description: SOS sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyAlert"
  /emergency/{id}/resolve:
    patch:
      operationId: resolveEmergency
      tags: [emergency]
      summary: Resolve an emergency alert
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyAlert"

  # DASHBOARD
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterInput:
      type: object
      required: [passportNo, fullNameAr, nationality, phone, password]
      properties:
        passportNo:
          type: string
        fullNameAr:
          type: string
        fullNameEn:
          type: ["string", "null"]
        nationality:
          type: string
          minLength: 2
          maxLength: 2
        phone:
          type: string
        password:
          type: string
          minLength: 6
        tentZone:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, passportNo, fullNameAr, nationality, phone, createdAt]
      properties:
        id:
          type: string
          format: uuid
        passportNo:
          type: string
        fullNameAr:
          type: string
        fullNameEn:
          type: ["string", "null"]
        nationality:
          type: string
        phone:
          type: string
        tentZone:
          type: ["string", "null"]
        groupId:
          type: ["string", "null"]
          format: uuid
        emergencyContact:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # GROUPS
    GroupInput:
      type: object
      required: [nameAr]
      properties:
        nameAr:
          type: string
        agency:
          type: ["string", "null"]
        maxMembers:
          type: ["integer", "null"]
    JoinGroupInput:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string
    Group:
      type: object
      required: [id, nameAr, inviteCode, createdAt]
      properties:
        id:
          type: string
          format: uuid
        nameAr:
          type: string
        agency:
          type: ["string", "null"]
        leaderId:
          type: ["string", "null"]
          format: uuid
        inviteCode:
          type: string
        maxMembers:
          type: integer
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    GroupMember:
      type: object
      required: [id, fullNameAr, phone]
      properties:
        id:
          type: string
          format: uuid
        fullNameAr:
          type: string
        fullNameEn:
          type: ["string", "null"]
        phone:
          type: string
        tentZone:
          type: ["string", "null"]
        lastLat:
          type: ["number", "null"]
        lastLng:
          type: ["number", "null"]
        lastSeenAt:
          type: ["string", "null"]
          format: date-time

    # LOCATIONS
    LocationPoint:
      type: object
      required: [latitude, longitude, recordedAt]
      properties:
        latitude:
          type: number
        longitude:
          type: number
        accuracy:
          type: ["number", "null"]
        zone:
          type: ["string", "null"]
        recordedAt:
          type: string
          format: date-time
        isOfflineSync:
          type: boolean
    LocationBatch:
      type: object
      required: [locations]
      properties:
        locations:
          type: array
          items:
            $ref: "#/components/schemas/LocationPoint"
    SyncResult:
      type: object
      required: [synced]
      properties:
        synced:
          type: integer
        message:
          type: string
...
    ManasikItem:
      type: object
      required: [key, titleAr, titleEn, order, status]
      properties:
        key:
          type: string
        titleAr:
          type: string
        titleEn:
          type: string
        descriptionAr:
          type: ["string", "null"]
        order:
          type: integer
        day:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed]
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
    ManasikProgressUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed]

    # PLACES
...
    SosInput:
      type: object
      properties:
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        alertType:
          type: string
    EmergencyAlert:
      type: object
      required: [id, userId, alertType, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        alertType:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, resolved]
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time

    # DASHBOARD
...