# lib/api-zod/src/generated/types/locationBatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hajj Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { LocationPoint } from "./locationPoint";

export interface LocationBatch {
  locations: LocationPoint[];
}