# lib/api-zod/src/generated/types/manasikProgressUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hajj Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { ManasikProgressUpdateStatus } from "./manasikProgressUpdateStatus";

export interface ManasikProgressUpdate {
  status: ManasikProgressUpdateStatus;
}