# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hajj Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  passportNo: string;
  fullNameAr: string;
  /** @nullable */
  fullNameEn?: string | null;
  nationality: string;
  phone: string;
  /** @nullable */
  tentZone?: string | null;
  /** @nullable */
  groupId?: string | null;
  /** @nullable */
  emergencyContact?: string | null;
  createdAt: Date;
}